Move ANSI tests to separate file
