+++ test/smoke/src/main.ts
@@ -45,8 +45,9 @@ import { setup as setupHelpTest } from './areas/positron/help/help.test';
 import { setup as setupClipboardTest } from './areas/positron/console/consoleClipboard.test';
 import { setup as setupLayoutTest } from './areas/positron/layouts/layouts.test';
 import { setup as setupTopActionBarTest } from './areas/positron/top-action-bar/top-action-bar.test';
-import { setup as setupNotebookVariablesTest} from './areas/positron/variables/notebookVariables.test';
+import { setup as setupNotebookVariablesTest } from './areas/positron/variables/notebookVariables.test';
 import { setup as setupConsoleInputTest } from './areas/positron/console/consoleInput.test';
+import { setup as setupConsoleANSITest } from './areas/positron/console/consoleANSI.test';
 import { setup as setupConsoleOutputLogTest } from './areas/positron/output/consoleOutputLog.test';
 import { setup as setupBasicRMarkdownTest } from './areas/positron/rmarkdown/rmarkdown.test';
 // --- End Positron ---
@@ -452,6 +453,7 @@ describe(`VSCode Smoke Tests (${opts.web ? 'Web' : 'Electron'})`, () => {
 	setupLayoutTest(logger);
 	setupNotebookVariablesTest(logger);
 	setupConsoleInputTest(logger);
+	setupConsoleANSITest(logger);
 	setupConsoleOutputLogTest(logger);
 	setupBasicRMarkdownTest(logger);
 	// --- End Positron ---
